Some Changes

+++ DeePore.py
@@ -243,8 +243,8 @@ def gener(batch_size,Data,List,MIN,MAX):
         number_of_batches = int(samples_per_epoch/batch_size)
         counter=0
         while 1:
-            t1=f['X'][np.sort(List[batch_size*counter:batch_size*(counter+1)]),...]
-            t2=f['Y'][np.sort(List[batch_size*counter:batch_size*(counter+1)]),...]
+            t1=f['X'][np.int32(np.sort(List[batch_size*counter:batch_size*(counter+1)])),...]
+            t2=f['Y'][np.int32(np.sort(List[batch_size*counter:batch_size*(counter+1)])),...]
             X_batch=t1.astype('float32')/128
             y_batch=t2.astype('float32')
             y_batch=np.reshape(y_batch,(y_batch.shape[0],y_batch.shape[1]))
